Handle commas in parameter's templated types.
The split_params() function has now been added.
This acts similar to calling `.split(',')` on the parameter text,
however it only splits on commas which are not embedded within a
set of angle brackets (<>). Other bracket types are not handled as
they are separable by a code observer taking advantage of the scopes
denoted by code_walk().